pyproject.toml
[project]
name = "localllm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pydantic-settings>=2.6.1",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "python3-discogs-client>=2.7.1",
    "spotipy>=2.24.0",
    "sqlmodel>=0.0.22",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",
]

[project.scripts]
localllm = "localllm:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "coverage>=7.6.9",
    "pre-commit>=4.0.1",
    "ruff-lsp>=0.0.59",
    "ruff>=0.7.4",
]
test = [
    "anyio>=4.7.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "pytest>=8.3.4",
]

[tool.pytest.ini_options]
addopts = "-vvv"
testpaths = "tests"

[tool.ruff]
extend-exclude = [
    "__pycache__",
    "build",
    "dist",
]
target-version = "py312"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
extend-select = [
    "C4",
    "D200",
    "D201",
    "D204",
    "D205",
    "D206",
    "D210",
    "D211",
    "D213",
    "D300",
    "D400",
    "D402",
    "D403",
    "D404",
    "D419",
    "E",
    "F",
    "G010",
    "I",
    "INP001",
    "N805",
    "PERF101",
    "PERF102",
    "PERF401",
    "PERF402",
    "PGH004",
    "PGH005",
    "PIE794",
    "PIE796",
    "PIE807",
    "PIE810",
    "RET502",
    "RET503",
    "RET504",
    "RET505",
    "RUF015",
    "RUF100",
    "S101",
    "T20",
    "UP",
    "W",
]

[tool.ruff.lint.per-file-ignores]
"tests/*.py" = ["S101"]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "$version"
version_scheme = "semver"
version_provider = "pep621"
update_changelog_on_bump = true
major_version_zero = true
